Add keyboard shortcuts for sidebar
Keyboard code in sidebar.tsx similar as in controls-bar.tsx

Pressing the tilde key (`) goes back a tab, and pressing (1) goes to the
next tab. Currently, it only goes through the following tabs:

* Game
* Queue
* Runner
* Profiler

The Map-Editor is skipped because it completely stops the current game
and is unlikely to be used by most contestants. Help is skipped because
it is unlikely to be used frequently in the development process for each
contestant.