Fix double-click bug in controls-bar.tsx
If the competitor clicked on a button on the controls-bar right before
using the pause shortcut (which is assigned to space), the accessibility
features of the browser would interpret the space as an interaction with
the most recently selected element, causing a double-click. Delect ONLY
button elements when a key shortcut is used by using HTMLElement.blur().